created
The provided code is a simple web form created using HTML, CSS, JavaScript, and jQuery. It demonstrates how to use AJAX to fetch data from an external API (ReqRes.in) and dynamically populate form fields with the retrieved data. The form includes input fields for the user's first name and last name, along with a static image. Upon clicking the "Submit" button, an AJAX request is made to the API to fetch user data. The retrieved first name ,last name and image are then populated into the corresponding input fields. This project serves as a basic example of using AJAX to interact with APIs and update form elements dynamically.